fix serial comunication

+++ src/PIDController.cpp
@@ -1,6 +1,6 @@
 #include "PIDController.h"
 
-PIDController::PIDController(double Kp_, double Ki_, double Kd_) {
+PIDController::PIDController(float Kp_, float Ki_, float Kd_) {
     Kp = Kp_;
     Ki = Ki_;
     Kd = Kd_;
@@ -8,10 +8,10 @@ PIDController::PIDController(double Kp_, double Ki_, double Kd_) {
     _i = 0;
 }
 
-double PIDController::calc(double x, double xd, double dt) {
-    double e = xd - x;
+float PIDController::calc(float x, float xd, float dt) {
+    float e = xd - x;
     _i += dt * e;
-    double d = (e - _e_prev) / dt;
+    float d = (e - _e_prev) / dt;
     _e_prev = e;
 
     return Kp * e + Ki * _i + Kd * d;
